requirements.txt
APScheduler==3.6.3
awsebcli==3.20.3
backports.zoneinfo==0.2.1
boto3==1.24.94
botocore==1.23.54
cachetools==4.2.2
cement==2.8.2
certifi==2022.9.24
charset-normalizer==2.0.12
colorama==0.4.3
flake8==5.0.4
future==0.16.0
idna==3.4
jmespath==0.10.0
mccabe==0.7.0
pathspec==0.9.0
pycodestyle==2.9.1
pyflakes==2.5.0
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.6.0
semantic-version==2.8.5
six==1.14.0
termcolor==1.1.0
tornado==6.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.12
wcwidth==0.1.9
